requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
Deprecated==1.2.13
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.0
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.3
pycparser==2.21
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
